New feature: food system requires mercs to eat and drink in order to survive - food and water levels are lowered every hour, eat food to survive - not doing so leads to various penalties and ultimately death - added food items for that purpose. Food can degrade over time. Rotten food poisons. - Mercs automatically consume food if hungry/thirsty and food is in their inventories - for more info, see http://www.bears-pit.com/board/ubbthreads.php/topics/307396/Re_Mercs_need_food_and_water_t.html#Post307396 - this feature requires new STI files from the GameDir trunk.
- added a bunch of filler variable to the soldier type, so hopefully savegame compatibility can be maintained easier in the future.

WARNING: This will break savegame compatibility

git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@5411 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ Tactical/Weapons.cpp
@@ -11285,7 +11285,7 @@ void GunIncreaseHeat( OBJECTTYPE *pObj )
 
 	  FLOAT singleshottemperature = GetSingleShotTemperature( pObj );						// ... get temperature rise ...
 
-	  FLOAT newguntemperature = min(guntemperature + singleshottemperature, (FLOAT)(OVERHEATING_MAX_TEMPERATURE) );					// ... calculate new temperature ...
+	  FLOAT newguntemperature = min(guntemperature + singleshottemperature, OVERHEATING_MAX_TEMPERATURE );					// ... calculate new temperature ...
 
 	  (*pObj)[0]->data.bTemperature = newguntemperature;									// ... apply new temperature
 
@@ -11353,7 +11353,7 @@ FLOAT   GetGunOverheatJamPercentage( OBJECTTYPE * pObj )
 
 FLOAT GetOverheatJamThreshold( OBJECTTYPE *pObj )
 {
-	FLOAT jamthreshold = (FLOAT) (OVERHEATING_MAX_TEMPERATURE / 4.0);
+	FLOAT jamthreshold = OVERHEATING_MAX_TEMPERATURE / 4.0f;
 
 	if ( Item[pObj->usItem].usItemClass & (IC_GUN|IC_LAUNCHER) )
 	{
@@ -11379,7 +11379,7 @@ FLOAT GetOverheatJamThreshold( OBJECTTYPE *pObj )
 
 FLOAT GetOverheatDamageThreshold( OBJECTTYPE *pObj )
 {
-	FLOAT damagethreshold = (FLOAT) (OVERHEATING_MAX_TEMPERATURE / 4.0);
+	FLOAT damagethreshold = OVERHEATING_MAX_TEMPERATURE / 4.0f;
 
 	if ( Item[pObj->usItem].usItemClass & (IC_GUN|IC_LAUNCHER) )
 	{
